% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sakura.R
\docType{package}
\name{sakura-package}
\alias{sakura}
\alias{sakura-package}
\title{sakura: Extension to R Serialization}
\description{
Extends the functionality of R serialization by augmenting the built-in
reference hook system. This enhanced implementation allows an integrated
single-pass operation that combines R serialization with third-party
serialization methods. Facilitates the serialization of even complex R
objects, which contain non-system reference objects, such as those accessed
via external pointers, to enable their use in parallel and distributed
computing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://shikokuchuo.net/sakura/}
  \item \url{https://github.com/shikokuchuo/sakura/}
  \item Report bugs at \url{https://github.com/shikokuchuo/sakura/issues}
}

}
\author{
\strong{Maintainer}: Charlie Gao \email{charlie.gao@shikokuchuo.net} (\href{https://orcid.org/0000-0002-0750-061X}{ORCID})

Authors:
\itemize{
  \item Travers Ching \email{traversc@gmail.com}
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
  \item Hibiki AI Limited [copyright holder]
}

}
