% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sakura.R
\docType{package}
\encoding{UTF-8}
\name{sakura-package}
\alias{sakura}
\alias{sakura-package}
\title{sakura: Extension to R Serialization}
\description{
Exposes the 'refhook' functionality of R serialization for alternative
serialization of non-system reference objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://shikokuchuo.net/sakura/}
  \item \url{https://github.com/shikokuchuo/sakura/}
  \item Report bugs at \url{https://github.com/shikokuchuo/sakura/issues}
}

}
\author{
Charlie Gao \email{charlie.gao@shikokuchuo.net}
  (\href{https://orcid.org/0000-0002-0750-061X}{ORCID})
}
