% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sakura.R
\name{serial_config}
\alias{serial_config}
\title{Create Serialization Configuration}
\usage{
serial_config(class, sfunc, ufunc)
}
\arguments{
\item{class}{a character string (or vector) of the class of object custom
serialization functions are applied to, e.g. \code{'ArrowTabular'} or
\code{c('torch_tensor', 'ArrowTabular')}.}

\item{sfunc}{a function (or list of functions) that accepts a reference
object inheriting from \code{class} and returns a raw vector.}

\item{ufunc}{a function (or list of functions) that accepts a raw vector and
returns a reference object.}
}
\value{
A pairlist comprising the configuration. This may be provided to the
\code{hook} argument of \code{\link[=serialize]{serialize()}} and \code{\link[=unserialize]{unserialize()}}.
}
\description{
Returns a serialization configuration for custom serialization and
unserialization of non-system reference objects, using the 'refhook' system
of R native serialization. This allows their use across different R sessions.
}
\examples{
serial_config("test_class", base::serialize, base::unserialize)

serial_config(
  c("class_one", "class_two"),
  list(base::serialize, base::serialize),
  list(base::unserialize, base::unserialize)
)

}
