% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sakura.R
\name{serialize}
\alias{serialize}
\alias{unserialize}
\title{Serialize}
\usage{
serialize(x, hook = NULL)

unserialize(x, hook = NULL)
}
\arguments{
\item{x}{an object.}

\item{hook}{[default NULL] optionally, a configuration returned by
\code{\link{serial_config}}.}
}
\value{
For serialize: a raw vector. For unserialize: the unserialized object.
}
\description{
An extension of R native serialization using the 'refhook' system for custom
serialization and unserialization of non-system reference objects.
}
\examples{
vec <- serialize(data.frame())
vec
unserialize(vec)

}
